+++ loqusdb/build_models/__init__.py
@@ -1,2 +1,3 @@
 from .case import build_case
 from .variant import build_variant
+from .profile_variant import build_profile_variant