+++ loqusdb/exceptions/__init__.py
@@ -1,2 +1,3 @@
 from .case import CaseError
 from .vcf import VcfError
+from .profile import ProfileError