+++ loqusdb/models/__init__.py
@@ -1,3 +1,4 @@
 from .case import (Case, Individual)
 from .variant import Variant
 from .identity import Identity
+from .profile_variant import ProfileVariant