+++ loqusdb/utils/load.py
@@ -16,14 +16,17 @@ from pprint import pprint as pp
 from .vcf import (get_vcf, check_vcf)
 from .case import (get_case, update_case)
 from .delete import delete
-from loqusdb.build_models import (build_case, build_variant)
+from .profiling import (get_profiles, profile_match)
+from loqusdb.build_models import (build_case, build_variant, build_profile_variant)
 from loqusdb.exceptions import (CaseError, VcfError)
+
 
 LOG = logging.getLogger(__name__)
 
 def load_database(adapter, variant_file=None, sv_file=None, family_file=None,
                   family_type='ped', skip_case_id=False, gq_treshold=None,
-                  case_id=None, max_window = 3000):
+                  case_id=None, max_window = 3000, profile_file=None,
+                  hard_threshold=0.95, soft_threshold=0.9):
     """Load the database with a case and its variants
 
     Args:
@@ -36,6 +39,9 @@ def load_database(adapter, variant_file=None, sv_file=None, family_file=None,
           gq_treshold(int): If only quality variants should be considered
           case_id(str): If different case id than the one in family file should be used
           max_window(int): Specify the max size for sv windows
+          check_profile(bool): Does profile check if True
+          hard_threshold(float): Rejects load if hamming distance above this is found
+          soft_threshold(float): Stores similar samples if hamming distance above this is found
 
     Returns:
           nr_inserted(int)
@@ -59,6 +65,16 @@ def load_database(adapter, variant_file=None, sv_file=None, family_file=None,
         nr_sv_variants = vcf_info['nr_variants']
         vcf_files.append(sv_file)
         sv_individuals = vcf_info['individuals']
+
+    profiles = None
+    matches = None
+    if profile_file:
+        profiles = get_profiles(adapter, profile_file)
+        ###Check if any profile already exists
+        matches = profile_match(adapter,
+                                profiles,
+                                hard_threshold=hard_threshold,
+                                soft_threshold=soft_threshold)
 
     # If a gq treshold is used the variants needs to have GQ
     for _vcf_file in vcf_files:
@@ -94,6 +110,9 @@ def load_database(adapter, variant_file=None, sv_file=None, family_file=None,
         vcf_sv_path=sv_file,
         sv_individuals=sv_individuals,
         nr_sv_variants=nr_sv_variants,
+        profiles=profiles,
+        matches=matches,
+        profile_path=profile_file
     )
     # Build and load a new case, or update an existing one
     load_case(
@@ -185,9 +204,9 @@ def load_variants(adapter, vcf_obj, case_obj, skip_case_id=False, gq_treshold=No
         case_id = None
     # Loop over the variants in the vcf
     with click.progressbar(vcf_obj, label="Inserting variants",length=nr_variants) as bar:
-        
+
         variants = (build_variant(variant,case_obj,case_id, gq_treshold) for variant in bar)
-        
+
     if variant_type == 'sv':
         for sv_variant in variants:
             if not sv_variant:
@@ -201,3 +220,29 @@ def load_variants(adapter, vcf_obj, case_obj, skip_case_id=False, gq_treshold=No
     LOG.info("Inserted %s variants of type %s", nr_inserted, variant_type)
 
     return nr_inserted
+
+def load_profile_variants(adapter, variant_file):
+
+    """
+
+        Loads variants used for profiling
+
+        Args:
+            adapter (loqusdb.plugins.Adapter): initialized plugin
+            variant_file(str): Path to variant file
+
+
+    """
+
+    vcf_info = check_vcf(variant_file)
+    nr_variants = vcf_info['nr_variants']
+    variant_type = vcf_info['variant_type']
+
+    if variant_type != 'snv':
+        LOG.critical('Variants used for profiling must be SNVs only')
+        raise VcfError
+
+    vcf = get_vcf(variant_file)
+
+    profile_variants = [build_profile_variant(variant) for variant in vcf]
+    adapter.add_profile_variants(profile_variants)