+++ loqusdb/utils/vcf.py
@@ -7,35 +7,35 @@ from loqusdb.exceptions import VcfError
 from loqusdb.build_models.variant import get_variant_id
 
 LOG = logging.getLogger(__name__)
-VALID_ENDINGS = ['.vcf', '.gz']
+VALID_ENDINGS = ['.vcf', '.gz', '.bcf']
 
 def add_headers(vcf_obj, nr_cases=None, sv=False):
     """Add loqus specific information to a VCF header
-    
+
     Args:
         vcf_obj(cyvcf2.VCF)
     """
 
     vcf_obj.add_info_to_header(
         {
-            'ID':"Obs", 
-            'Number': '1', 
-            'Type': 'Integer', 
+            'ID':"Obs",
+            'Number': '1',
+            'Type': 'Integer',
             'Description': "The number of observations for the variant"}
     )
     if not sv:
         vcf_obj.add_info_to_header(
             {
-                'ID':"Hom", 
-                'Number': '1', 
-                'Type': 'Integer', 
+                'ID':"Hom",
+                'Number': '1',
+                'Type': 'Integer',
                 'Description': "The number of observed homozygotes"}
         )
         vcf_obj.add_info_to_header(
             {
-                'ID':"Hem", 
-                'Number': '1', 
-                'Type': 'Integer', 
+                'ID':"Hem",
+                'Number': '1',
+                'Type': 'Integer',
                 'Description': "The number of observed hemizygotes"}
         )
     if nr_cases:
@@ -43,43 +43,43 @@ def add_headers(vcf_obj, nr_cases=None, sv=False):
         vcf_obj.add_to_header(case_header)
     # head.add_version_tracking("loqusdb", version, datetime.now().strftime("%Y-%m-%d %H:%M"))
     return
-    
-    
+
+
 
 def get_file_handle(file_path):
     """Return cyvcf2 VCF object
-    
+
     Args:
         file_path(str)
-    
+
     Returns:
         vcf_obj(cyvcf2.VCF)
     """
     LOG.debug("Check if file end is correct")
-    
+
     if not os.path.exists(file_path):
         raise IOError("No such file:{0}".format(file_path))
-    
+
     if not os.path.splitext(file_path)[-1] in VALID_ENDINGS:
         raise IOError("Not a valid vcf file name: {}".format(file_path))
-    
+
     vcf_obj = VCF(file_path)
-    
+
     return vcf_obj
 
 def get_vcf(file_path):
     """Yield variants from a vcf file
-    
+
     Args:
         file_path(str)
-    
+
     Yields:
         vcf_obj(cyvcf2.VCF): An iterable with cyvcf2.Variant
 
     """
 
     vcf_obj = get_file_handle(file_path)
-    
+
     return vcf_obj
 
 def check_sorting(previous_chrom, previous_pos, current_chrom, current_pos):
@@ -94,9 +94,9 @@ def check_vcf(vcf_path, expected_type='snv'):
         expected_type(str): 'sv' or 'snv'
 
     Returns:
-        vcf_info(dict): dict like 
+        vcf_info(dict): dict like
         {
-            'nr_variants':<INT>, 
+            'nr_variants':<INT>,
             'variant_type': <STR> in ['snv', 'sv'],
             'individuals': <LIST> individual positions in file
         }
@@ -126,7 +126,7 @@ def check_vcf(vcf_path, expected_type='snv'):
 
         current_chrom = variant.CHROM
         current_pos = variant.POS
-        
+
         # We start with a simple id that can be used by SV:s
         variant_id = "{0}_{1}".format(current_chrom, current_pos)
         # For SNVs we can create a proper variant id with chrom_pos_ref_alt
@@ -146,7 +146,7 @@ def check_vcf(vcf_path, expected_type='snv'):
             previous_pos = current_pos
             posititon_variants = set([variant_id])
             continue
-        
+
         if variant_type == 'snv':
             # Check if variant is unique
             if current_pos == previous_pos:
@@ -170,7 +170,7 @@ def check_vcf(vcf_path, expected_type='snv'):
     LOG.info("Vcf file %s looks fine", vcf_path)
     LOG.info("Nr of variants in vcf: {0}".format(nr_variants))
     LOG.info("Type of variants in vcf: {0}".format(variant_type))
-    
+
     vcf_info = {
         'nr_variants': nr_variants,
         'variant_type': variant_type,